Move jdbc stuff to proper namespace
