+++ lib/active_record/sqlserver_base.rb
@@ -6,7 +6,7 @@ module ActiveRecord
       mode = config[:mode].to_s.downcase.underscore.to_sym
       case mode
       when :jdbc
-        require 'activerecord_sqlserver_adapter/jdbc'
+        require 'active_record/connection_adapters/sqlserver/jdbc'
       when :dblib
         require 'tiny_tds'
       else
